+++ examples/Basic/source/main.cpp
@@ -1,0 +1,20 @@
+#include <Small/All.hpp>
+
+#include <MyScene.hpp>
+
+int main(int argc, char** argv)
+{
+    const sgl::Settings settings = {};
+    sgl::Application app(settings);
+
+    app.setTitle("Hello world!")
+        .setVSync(false)
+        .setFPSLimit(60)
+        .setDebug(true);
+    
+    app.add<MyScene>(12);
+
+    app.run();
+
+    return 0;
+}