+++ include/Small/All.hpp
@@ -1,4 +1,17 @@
+/**
+ * @file All.hpp
+ * @author Alexandre Plateau (lexplt.dev@gmail.com)
+ * @brief Include the needed files to start a project using the library
+ * @version 0.1
+ * @date 2020-04-05
+ * 
+ * @copyright Copyright (c) 2020
+ * 
+ */
+
 #ifndef sgl_small_all
 #define sgl_small_all
+
+#include <sgl/Core/Application.hpp>
 
 #endif