+++ include/Small/Core/SceneManager.hpp
@@ -1,0 +1,122 @@
+/**
+ * @file SceneManager.hpp
+ * @author Alexandre Plateau (lexplt.dev@gmail.com)
+ * @brief Define the scene manager (creating, handling and destroying them)
+ * @version 0.1
+ * @date 2020-04-05
+ * 
+ * @copyright Copyright (c) 2020
+ * 
+ */
+
+#ifndef sgl_core_scenemanager
+#define sgl_core_scenemanager
+
+#include <vector>
+#include <memory>
+#include <SFML/Graphics.hpp>
+#include <SFML/System/Clock.hpp>
+#include <SFML/Window/Event.hpp>
+#include <Small/Core/Scene.hpp>
+
+namespace sgl
+{
+    /**
+     * @brief Scene manager, in charge of creating, handling and destroying scenes
+     * 
+     */
+    class SceneManager
+    {
+    public:
+        /**
+         * @brief Construct a new Scene Manager object
+         * 
+         */
+        SceneManager();
+
+        /**
+         * @brief Destroy the Scene Manager object
+         * 
+         */
+        ~SceneManager();
+
+        /**
+         * @brief Return the identifier of the current scene
+         * 
+         * @return int 
+         */
+        inline int current() const;
+
+        /**
+         * @brief Remove a scene by searching it from its identifier
+         * 
+         * @param id 
+         * @return true On success
+         * @return false If the scene couldn't be found
+         */
+        bool remove(int id);
+
+        /**
+         * @brief Return a scene by its identifier
+         * 
+         * @param id 
+         * @return Scene* 
+         */
+        inline Scene* operator[](int id);
+
+        /**
+         * @brief Set the Current scene
+         * 
+         * @param id 
+         * @return SceneManager& 
+         */
+        SceneManager& setCurrent(int id);
+
+        /**
+         * @brief Dispatch the event to the active scenes
+         * 
+         * @param event 
+         * @return SceneManager& 
+         */
+        SceneManager& handleEvent(const sf::Event& event);
+
+        /**
+         * @brief Update the active scenes
+         * 
+         * @param dt 
+         * @return SceneManager& 
+         */
+        SceneManager& update(const sf::Time dt);
+
+        /**
+         * @brief Render the active scenes
+         * 
+         * @param screen 
+         * @return SceneManager& 
+         */
+        SceneManager& render(sf::RenderTarget& screen);
+
+        /**
+         * @brief Register a scene and return its identifier
+         * 
+         * @tparam S The type of the scene
+         * @tparam Args 
+         * @param args Arguments for the constructor of the scene
+         * @return int Identifier of the scene
+         */
+        template <typename S, typename... Args>
+        int add(Args&&... args)
+        {
+            m_scenes.push_back(std::make_unique<S>(m_scenes.size(), std::forward<Args>(args)...));
+            return static_cast<int>(m_scenes.size()) - 1;
+        }
+    
+    private:
+        std::vector<std::unique_ptr<Scene>> m_scenes;
+        int m_current;
+    };
+}
+
+#include "SceneManager.inl"
+
+#endif